Refactoring navbar code.
