+++ src/sass/components/_pill.scss
@@ -8,7 +8,6 @@
   font-size: 1rem;
   cursor: pointer;
   margin: $small;
-  max-width: 48rem;
   &--active {
     background-image: $primary-gradient;
     box-shadow: $glow;
